requirements.txt
fastapi==0.68.0
uvicorn==0.15
ffmpeg-python==0.2.0
streaming-form-data==1.12.0
python-multipart==0.0.6